Add linter to automated test script
Previously there was a problem where a syntax error in the code run by
Nightwatch would not cause an overall non-zero exit code.

As a quick and dirty fix, this commit ensures that the linter also runs
during the test process. Each time, it will run both the linter and the
test suite, but will exit with an error code of 1 if either do not exit
cleanly with an exit code of 0.